split api_generator.zig into five focused modules
